make gate compile also with geant 4.10.01
There were some API changes in 4.10.01 that caused the compilation of
GATE to fail:
* the GetIon() functionality was separated from G4ParticleTable, if you
  want to call it with 4.10.01 then you need to insert GetIonTable()
* particle definitions are supposed to be constant, and 4.10.01 is
  enforcing this more consistently than in older versions.

I solved these two issues now by adding the geant major and minor
version numbers to the compiler flags, and making parts of the code
version dependent.  Maybe there are more elegant solutions.

I think that the use of 'const_cast' should be avoided as much as
possible.  With 4.10.01 it can indeed be eliminated from
GateBinaryCascade altogether, with 4.9.6.p04 I did not see an easy way
to avoid it.

NOTE that these patches *only* makes sure that the code actually
compiles (on my RHEL6 box). There may be changes in geant 4.10 that
cause GATE to run incorrectly. For instance 4.10 has an option to run
multithreaded; I have no idea if GATE is ready to run multithreaded (I
have practically no experience with multithreaded programming).