Make the nvcc (cuda) compile also work with C++11
When std C++11 is requested then NVCC needs to know about that too, so I
added a flag to the NVCC flags when the user requests C++11. I checked
that this works with CUDA 8.0 on an Ubuntu 14.04 workstation with an
NVidia Quadro 600 graphics card. I do not have the resources to check
this with arbitrary combinations of OS and hardware, so it would be good
if others can test this too.

One thing I noticed is that the "compute capabilities" are hardcoded to
some values that were current many years ago. This needs to be updated
in a smart way, such that we only try to compile for relevant compute
capabilities, such as the ones that are supported by the installed
CUDA version, or even only the one needed for the GPU that is available
on the build machine.