sanitize HU-to-materials table range checking
If some HU values are too large, use material corresponding to max HU value in the table.
If some HU values are too low, use material corresponding to min HU value in the table.
If too many voxels have out-of-range HU values, exit with an error.

Challenge: what to do with 'regionalized' images that add an extra layer of voxels
and fill it with the minimum value of the image minus one?
Right now I am just ignoring them for the decision about whether there are "too many".
Ideal solution is suggested somewhere in the comments: the 'outside value' should
actually NOT be computed from the image itself, but be taken as the minimum value
from the materials tables. However, it's not so clear how/if that can work with the
different kinds of tables.